requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-shopping-cart @ file:///Users/ankitagrawal/Downloads/django_shopping_cart-0.1-py3-none-any.whl
Django-Verify-Email==1.0.6
djangorestframework==3.12.4
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.2
pycparser==2.20
PyJWT==2.0.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.4
